feat: move Purple enum and related implementations to a separate module and implement `to_rgb` and `to_hex_triplet`
